(Bug Fix) Bug: int value cast to float and then passed to Math.round

+++ branches/GWT-2.0/src/org/gwt/mosaic/core/client/util/AbstractUnitConverter.java
@@ -228,7 +228,7 @@ public abstract class AbstractUnitConverter implements UnitConverter {
    * @return the given Points as pixels
    */
   protected final int pointAsPixel(int pt, int dpi) {
-    return Math.round(dpi * pt / DTP_RESOLUTION);
+    return (int) Math.round((double)(dpi * pt) / DTP_RESOLUTION);
   }
 
   /**
